Updating the folder structure for analytics to add ACPAnalyticsFolder
