--- ACPAnalytics/Assets/Plugins/ACPAnalytics.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 672a091b638a5476c8470b788ca106e0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
