+++ ACPAnalytics/Assets/Plugins/ACPAnalytics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0eb587bf01d694897a513ae0f7f36b59
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
