+++ ACPAnalytics/Assets/Plugins/iOS/ACPAnalytics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 09962c74f59664f88aaaa26440c84735
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
