+++ Makefile
@@ -9,6 +9,9 @@ BIN_DIR=$(ROOT_DIR)/bin
 BUILD_DIR=$(BIN_DIR)/build_temp
 BUILD_PKG=ACPAnalytics.unitypackage
 ASSETS=Assets/Plugins
+PLUGIN_NAME=ACPAnalytics
+IOS_DIR=$(ASSETS)/iOS/$(PLUGIN_NAME)
+ANDROID_DIR=$(ASSETS)/Android/$(PLUGIN_NAME)
 
 # targets
 release: clean setup unity_build
@@ -28,7 +31,7 @@ unity_build:
 	$(UNITY_BIN) -batchmode -quit \
 	-logFile $(BUILD_DIR)/buildLog.log \
 	-projectPath $(PROJECT_DIR) \
-	-exportPackage $(ASSETS) ../$(BUILD_DIR)/$(MOBILE_DIR)/$(BUILD_PKG)
+	-exportPackage $(ASSETS)/$(PLUGIN_NAME) $(IOS_DIR) $(ANDROID_DIR) ../$(BUILD_DIR)/$(MOBILE_DIR)/$(BUILD_PKG)
 
 	@echo ""
 	@echo "######################################################################"
